# docs/swagger/swagger.yaml
definitions:
  utils.ErrResp:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: DB Error Message
        type: string
    type: object
  utils.GenCodeReq:
    properties:
      password:
        example: password
        type: string
      phone:
        example: "17777777777"
        type: string
      username:
        example: username
        type: string
    type: object
  utils.GenResetCodeReq:
    properties:
      username:
        example: username
        type: string
    type: object
  utils.Ping:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Pong
        type: string
    type: object
  utils.RegisterReq:
    properties:
      username:
        example: username
        type: string
      verifyCode:
        example: "1234"
        type: string
    type: object
  utils.ResetReq:
    properties:
      password:
        example: password
        type: string
      username:
        example: username
        type: string
      verifyCode:
        example: "1234"
        type: string
    type: object
  utils.StdResp:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: OK
        type: string
    type: object
  utils.VerifyResp:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: OK
        type: string
      status:
        example: Verified
        type: string
    type: object
host: fmt.fredliang.cn
info:
  contact:
    email: info@fredliang.cn
    name: Fred Liang
    url: https://blog.fredliang.cn
  description: This is a Family Tree API server.
  license:
    name: Mozilla 2.0
    url: https://github.com/fredliang44/family-tree/blob/master/LICENSE
  title: Family Tree API
  version: "1.0"
paths:
  /init_db:
    get:
      consumes:
      - application/json
      description: InitDB for unit test
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/utils.StdResp'
            type: object
      security:
      - ApiKeyAuth: []
      summary: InitDB
  /ping:
    get:
      consumes:
      - application/json
      description: Respond pong to ping request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Ping'
            type: object
      summary: HealthCheck
  /register:
    post:
      consumes:
      - application/json
      description: Register after verify sms code
      parameters:
      - description: Verify Register Code
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/utils.RegisterReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.VerifyResp'
            type: object
        "400":
          description: DB Error
          schema:
            $ref: '#/definitions/utils.ErrResp'
            type: object
      summary: Register
  /register_code:
    post:
      consumes:
      - application/json
      description: Generate register phone sms auth code
      parameters:
      - description: Gen Register Code
        in: body
        name: GenCode
        required: true
        schema:
          $ref: '#/definitions/utils.GenCodeReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StdResp'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResp'
            type: object
      summary: Gen Register Code
  /reset:
    post:
      consumes:
      - application/json
      description: Phrase reset password request
      parameters:
      - description: Verify Reset Code
        in: body
        name: ResetPassword
        required: true
        schema:
          $ref: '#/definitions/utils.ResetReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.VerifyResp'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResp'
            type: object
      summary: Reset Password
  /reset_password_code:
    post:
      consumes:
      - application/json
      description: Generate register phone sms auth code
      parameters:
      - description: Gen Reset Code
        in: body
        name: GenResetCode
        required: true
        schema:
          $ref: '#/definitions/utils.GenResetCodeReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StdResp'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResp'
            type: object
      summary: Gen Reset Code
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"